Simplified GuessWindow and moved it into the GUI class
Simplified repetitive generated code by using an array and for loops, will allow us to not have to make code specific to each button in the future.
Changed guess window class to a JFrame inside GUI class, now takes system theme like the rest of the GUI.
Left original version of GuessWindow as a .txt for reference incase I happened to break something.