Server and Client updates (mostly)
Fixed server address on gui. Added sendGuess method to the Client object and linked it to the associated submit button. Added usage of updatePlayerCount when a new client has been accepted. Fixed the updatePlayerCount method to set the text properly.  Added some of the handling for connecting to the server from the client gui.